+++ tests/spec/test_union.py
@@ -26,7 +26,7 @@ from thriftrw.idl import Parser
 from thriftrw.spec.union import UnionTypeSpec
 from thriftrw.spec.struct import FieldSpec
 from thriftrw.spec.reference import TypeReference
-from thriftrw.compile.exceptions import ThriftCompilerError
+from thriftrw.errors import ThriftCompilerError
 from thriftrw.spec import primitive as prim_spec
 from thriftrw.spec.list import ListTypeSpec
 from thriftrw.wire.ttype import TType