+++ CHANGELOG.rst
@@ -5,6 +5,7 @@ Releases
 ------------------
 
 - Core modules have been cythonized for additional performance improvements.
+- Export all custom exceptions in the ``thriftrw.errors`` module.
 
 
 0.4.1 (2015-10-12)